cleanup file for newznab volume
removed svn checkout and permissions. files will now be downloaded to /var/www/newznab on first run of the container.  Added a volume for /var/www/newznab and removed chown of the www dir. Also added 2 new ENV for SNVUSER and SVNPASS to make it easier to keep my code base in line with this github repo without accidentally leaking the password.