added newznab-setup.sh
Added newznab-setup.sh to webserver build dir. This file is added to the container image and is run on startup of the container. If the notouch file does not exist it will run svn and download the latest newznab. it also does the chown and chmod commands removed from the Dockerfile and instead of adding config.php it simple echos the example config to the www dir.